Put CORS config into SecurityConfig
