+++ backend/src/main/java/com/tcs/dhv/config/SecurityConfig.java
@@ -15,7 +15,12 @@ import org.springframework.security.oauth2.server.resource.web.BearerTokenAuthen
 import org.springframework.security.oauth2.server.resource.web.access.BearerTokenAccessDeniedHandler;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.csrf.CookieCsrfTokenRepository;
+import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+
+import java.util.Arrays;
+import java.util.List;
 
 import static org.springframework.security.config.Customizer.withDefaults;
 
@@ -25,37 +30,52 @@ import static org.springframework.security.config.Customizer.withDefaults;
 public class SecurityConfig {
 
     public static final String[] PUBLIC_ENDPOINTS = {
-        "/api/auth/**",
-        "/api/tracking/**"
+            "/api/auth/**",
+            "/api/tracking/**"
     };
+
+    @Bean
+    public CorsConfigurationSource corsConfigurationSource() {
+        final var corsConfiguration = new CorsConfiguration();
+        corsConfiguration.setAllowedOrigins(List.of("http://localhost:3000"));
+        corsConfiguration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+        corsConfiguration.setAllowedHeaders(List.of("*"));
+        corsConfiguration.setAllowCredentials(true);
+        corsConfiguration.setMaxAge(3600L);
+
+        final var source = new UrlBasedCorsConfigurationSource();
+        source.registerCorsConfiguration("/**", corsConfiguration);
+
+        return source;
+    }
 
     @Bean
     public SecurityFilterChain securityFilterChain(
             final HttpSecurity http,
             CorsConfigurationSource corsConfigurationSource) throws Exception {
         http
-            .cors(cors -> cors.configurationSource(corsConfigurationSource))
-            .csrf(csrf -> csrf
-                .csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse())
-                .ignoringRequestMatchers(PUBLIC_ENDPOINTS)
-            )
-            .authorizeHttpRequests(auth -> {
-                auth.requestMatchers(PUBLIC_ENDPOINTS).permitAll();
-                auth.anyRequest().authenticated();
-            })
-            .sessionManagement(SecurityConfig::getSessionManagementConfig)
-            .oauth2ResourceServer(oauth2 -> oauth2
-                .jwt(withDefaults())
-                .authenticationEntryPoint(new BearerTokenAuthenticationEntryPoint())
-                .accessDeniedHandler(new BearerTokenAccessDeniedHandler())
-            );
+                .cors(cors -> cors.configurationSource(corsConfigurationSource))
+                .csrf(csrf -> csrf
+                        .csrfTokenRepository(CookieCsrfTokenRepository.withHttpOnlyFalse())
+                        .ignoringRequestMatchers(PUBLIC_ENDPOINTS)
+                )
+                .authorizeHttpRequests(auth -> {
+                    auth.requestMatchers(PUBLIC_ENDPOINTS).permitAll();
+                    auth.anyRequest().authenticated();
+                })
+                .sessionManagement(SecurityConfig::getSessionManagementConfig)
+                .oauth2ResourceServer(oauth2 -> oauth2
+                        .jwt(withDefaults())
+                        .authenticationEntryPoint(new BearerTokenAuthenticationEntryPoint())
+                        .accessDeniedHandler(new BearerTokenAccessDeniedHandler())
+                );
 
         return http.build();
     }
 
     @Bean
     public AuthenticationManager authenticationManager(
-        final AuthenticationConfiguration configuration
+            final AuthenticationConfiguration configuration
     ) throws Exception {
         return configuration.getAuthenticationManager();
     }
@@ -66,7 +86,7 @@ public class SecurityConfig {
     }
 
     private static void getSessionManagementConfig(
-        final SessionManagementConfigurer<HttpSecurity> session
+            final SessionManagementConfigurer<HttpSecurity> session
     ) {
         session.sessionCreationPolicy(SessionCreationPolicy.STATELESS);
     }
